Avoid dependency import cycle
